Add `Cargo.lock` to `.gitignore` and remove it from the repo
This removes `Cargo.lock` from the repository.
This also adds `Cargo.lock` to `.gitignore` because we do not want to impose 'pinned dependency requirements' on downstream users of the library.

See also: [Why do binaries have Cargo.lock in version control, but not libraries?](https://doc.rust-lang.org/cargo/faq.html#why-do-binaries-have-cargolock-in-version-control-but-not-libraries) from the Cargo book.