[cargo-diet] Add an include directive to minimize the package size of `atspi`.
Crates that added `atspi` as a dependency, do not need all of the repository.
Adding an [include or exclude directive](https://doc.rust-lang.org/cargo/reference/manifest.html?highlight=include%20directive#the-exclude-and-include-fields) allows to minimizing the projects published size.

Thisa saved 27% or 120.3 KB in 26 files (of 447.5 KB and 75 files in entire crate)

Done with the help of [cargo-diet](https://crates.io/crates/cargo-diet)