Change `Event` validation on messages to `MessageType::Signal` test.
Signals are broadcasted Events. So, to test for events, it is sufficient to test if the `Message` is `MessageType::Signal`.
This test is likely a little bit cheaper than the string compare, but that is not the point.
Point is that we allow 5 more streams to come through `RegistryProxy`:

$ rg ' signal' src/events/ -trust | wc --lines
56

$ rg ' signal' src/ -trust | wc --lines
61

The difference is made up by these events:

src/registry.rs
31:    /// EventListenerDeregistered signal
35:    /// EventListenerRegistered signal

src/cache.rs
59:    /// AddAccessible signal
63:    /// RemoveAccessible signal

src/socket.rs
30:    /// Available signal