[WIP] small fixes
It appears one cannot match against `const zvariant::Signature`s in the pattern position because `PartialEq` is manually implemented instead of derived.
The workaround is trivial but not as satisfying.

The comparison happens in the match gaurds now.
Zbus maintainer and I will dig into the problem a bit more tomorrow.