[WIP] improves errors from proc-macro
If an error occured in the derived TryFrom<Arc<Message>> for AtspiEvent, we would only know there was an error.
This makes the derived impl such that - if the comparison between implied type member and the member on the message do not correspond, both are named in the error.

having two `Signature` names sharing the same signature is problematic when matching because:
	- You either get an unreacheable match-arm or,
	- Botha arms have additional tests.
This replaces `AVAILABLE` and `CACHE_REM` consts and get replaced by `ACCESSIBLE`.
Match arm on ACCESSIBLE now decides which `*Event` is construed.