Destructure `Event` to final type without need to know internal structure of `Event`.
Atspi separates events by bus-type signature and groups them accordingly in enum structures, which in the end are grouped by the `Event` enum.
This event tree allows `atspi` to emit a conveniently typed heterogeneous variety of events using the same stream `event::event_stream()`.

To some users, the way we structured these events may be of little interest and may obscure or muddle their goal of obtaining the final type.
This PR allows users to hide the nested enum structure altogether. The only thing they need to know is the event of interest.

The `let-else` allows users to bind by naming the type and calling its `try_from` implementation on the nested event,
 which succeeds or continues to the next event in the case of a mismatch.

```rust
    while let Some(Ok(ev)) = stream.next().await {
        let Ok(change)  = <StateChangedEvent>::try_from(ev) else { continue; };

	// Do interesting things with `change` the `StateChangedEvent`.
    }
```