[Proposed] Replace documentation example
This offers for consideration a replacement of the example ships this branch, for `event_stream()`

The proposed example is a working minimalistic, (more idiomatic) example.
It's sole goal is to showcase how `event_stream()` may be used.

The doc-test guards that the example keeps working as/if the library evolves.