Fix bug that made Convertable::*() always succeed.
This was because there are two generic types for the blocking and non-blocking functions respectively.
U being the input type, T being the output type.

To convert to a new trait type, the underlying DBus object must implement the interface for the output (return) type, T.
But, it was checking for the input type, U.