Proposed fix for issue #83
See also issue #83

Registryd and presumably Qt applications implementing the `Cache` interface,
return a legacy formatted type when method `GetItems` is called.

This adds:
- A `LegacyCacheItem` type to match the `((so)(so)(so)a(so)assusau)` signature.
- Add `get_legacy_items` to the `Cache` interface.
- Add support for `LegacyCacheItem` events (Qt) on the bus.
This is a breaking change, because the enum `CacheEvents` lacks the `#[non_exhaustive]` attribute.
This means that atspi users who match on `CacheEvents` will experience breakage. (but it is for a good cause)

The reproducer by @Ablu, when linked against this branch and `get_items` replaced with `get_legacy_items` no longer yields an error:

```Term
   Compiling atspi v0.16.0 (/atspi-issue-83/atspi)
   Compiling cache_sucks v0.1.0 (/home/luuk/code/cache_reproducer)
    Finished dev [unoptimized + debuginfo] target(s) in 6.49s
     Running `target/debug/cache_reproducer`
[src/main.rs:14] items = []
```

Receiving Qt-cache signals needs to be tested still, but I would not know of an application that emits any.