Fix assumprion in Cache event tests
An assumption crept in test, which proved false.
If a `Message` was constructed and sent, the assumption was that the first event we would receive would be that event.
This was only mostly true, but sometimes it was not.

This commit fixess that by looking for the intended event on a Timeout stream.

Known issue with this fix is that the signatures do not correspond, due to outer parentheses.