Cleanups and link to modified zbus
This cleans up the tests on Cache `AddAccessible` and `RemoveAccessible`
Unfortunately the signature tests currently still fail, but this is likely due to some overlooked detail in the `Message` construction.

The new `PartialEq` supports elision of outer parentheses.
The tests in the modified `PartialEq` for `Signature` in zvariant all pass, that gives some confidence.