[atspi-common] macro split-up
The macro `impl_event_conversions` expands or can expand to five conversions
dependingon its arguments and it is not apparent at the call site to which
the macro will expand.

This introduces five separate macros and documents each,
such that at the call site, one will know which arguments are required and
what it will expand to.

The objective is twofold:
- to improve readability by knowing what the macro is short-hand for,
- to improve debugability by leading to a single implementation per macro.

Comes at the cost of more verbosity.

The new macros:

1. `mpl_try_from_event_enum_for_inner_type`
2. `impl_try_from_event_for_user_facing_event_type`
3. `impl_from_user_facing_event_for_interface_event_enum`
4. `impl_from_user_facing_type_for_event_enum`
5. `impl_try_from_event_for_user_facing_type`

Preferably these should have names that reflect what these expand to
at glance and do so in a verbiage that is consistent between them.