[CI] Add test for unmarshalled `AddAccessible` `Cache` message bodies/.
It appears to be common for bus servers to send `AddAccessible` signals over the
bus with unmarshalled bodies.
This requires us to special-case parsing `AddAccessible` `Cache` events in
`TryFrom<&Message> for Event` by also checking for bodies with outer parentheses.

This adds a test that sends such a 'compensating' body.