Add mode-level docs to AccessibleProxy as well.
The trait is transformed into structs with methods.
It is useful to have some global information to go with the struct and its methods.

It may seem redundant to repeat this header but as module level docs are separated from these structs in docs, I think it is warranted to find it in both places.