fix(common): `expect`-message to compliance with recommended message style
In the beginning of my Rust journey I believed 'expect' had a weird name
and it perhaps have been named 'unexpect' instead.

That was until I learned of [the recommended message style](ttps://doc.rust-lang.org/stable/std/result/enum.Result.html#recommended-message-style).

- When an application fails, it helps to learn what assumption made the author
believe the unwrap would the unwrap would succeed.
- The practice encourages the author to think about their assumptions.

In this macro it is demonstrated well:

```diff
macro_rules! generic_event_test_case {
	($type:ty) => {
		#[test]
		fn generic_event_uses() {
			let struct_event = <$type>::default();
			assert_eq!(struct_event.path().as_str(), "/org/a11y/atspi/accessible/null");
			assert_eq!(struct_event.sender().as_str(), ":0.0");
			let item = struct_event.item.clone();
			let body = struct_event.body();
-			let build_struct = <$type>::build(item, body).expect("Could not build type from parts");
+			let build_struct = <$type>::build(item, body)
+				.expect("<$type as Default>'s parts should build a valid ObjectRef");
			assert_eq!(struct_event, build_struct);
		}
	};
}
```

In the old version we more or less stated the obvious, while the new version
might be more helpful when it does fail.

On `Result` types we should trust the error to help us understand which varialbe
is/are involved.