fixup try_into / try_clone handling of `Value`.
On line object.rs:135

We no longer return a default but panic with helpful message.

On line object.rs:235

In this conversion, there are only `Strings` and a single `OwnedValue`,
which, at the end is just handed over and does not need to be cloned.
We should never see that `expect()` message.