fixup expect messages to reflect why we think these will succeed
Also added some clarification and docs in macros.rs because it was
not immediately apparent what was going on.

In the "Should fail" tests I had long winded explanations too, but decided to
revert those for this PR.

In tests we panic'ed on a method `sender(unique_name)` that can only fail if
the provided name cannot be converted into a `UniqueName<'a>`.
Since we provide it with a valid, owned variant, this will always succeed.