Changes `Interface` to have IntoStaticStr and Display derived
- Remove consts as we cannot use these inside the attribute macros anyway.
  (likely because these are expanded after proc macros get run.)
- Make up for loss of consts by adding tests for each.

Note that this does not include strum::FromRepr because the representation
differs from what is used on the bus, this makes the internal representatation
meaningless to users. Instead I implemented FromStr, which makes more sense.

Because of internal representation and bus representations iffer,
we need custom Serialize and Deserialize implementations.

Note that the Deserializer received a small optimization. No longer is the
interface string required to be copied to stack before converting it to an
`Interface`, instead the `Deserialize` implementation now takes a reference to
the string and converts it.

This also expands tests to cover the changes.