Role received strum traits
`FromRepr`, `Display`, `IntoStaticStr` and `EnumIter`

I changed two variants that stood out odd:

`Role::Editbar` -> `Role::EditBar`
``Role::CHART` ->  `Role::Chart`

In the case of `EditBar`, this means it is now displayed as "edit bar" instead
of "editbar". If this raises concerns, please inform me why.

the `EnumIter` may be useful in tests.
That being said, this enum has a simple `u32` representation.
fSo checking whether each gets serialized and deserialized feels overly cautious
because `zvariant` is surely able to (de)serialize these.
In other words, it is not our concern to test that. We added no logic to
get this done.

Same goes for `&'static str` matching with the `Display` impl. It does not hurt
to test, but it is not our responsibility.