Change `from_message` to `try_from_message`.
This thanges `BusProperties::from_message` to `try_from_message` to better
hint at what it returns.

Also fixes a dependency problem Clippy found in crate atspi, where we tried
to re-export zbus (but failed) because we did not have the `zbus` feature
present in that crate. This add zbus as optional dependency to that crate.