Help WASM target compile
We cannot compile zbus in WASM context, so this branch's use of `Message` in
the `BusProperties::try_from_message` method is problematic.

Because `BusProperties` is used to blanket-implement some other traits, this
commit tries to be least intrusive and avoids having `try_from_message`
in the trait when zbus is not set as feature.

The obvious downside is that `BusProperties` now has two sets of behaviour,
depending on the "zbus" feature.
That being said, `try_from_message` is not relevant anyway without zbus support.