Add lazy, cheaply guarded proxy-conversions.
This introduces lazy and cheaply guarded proxy-conversions.

It adds `ObjectProxies` which extends `AccessibleProxy

We have an easy conversion from `ObjectRef` to `AccessibleProxy`

`AccessibleProxy::get_interfaces` gives insight into which interfaces are
unokenebted by the object.

`ObjectProxies::proxies` returns `Proxies`, a struct that can contain zero or
more proxies to the objects supported interfaces.

The proxies are created only when needed and the check against the `InterfaceSet`
is a cheap bitwise operation.