fix(bench): Remove relative path to allow `cargo bench or criterion`
This removes the path to the stored `Message`s.
```Rust
cargo bench --quiet
```

or
```
cargo criterion
```

now both work as expected.
The problem was that benc and criterion differ in opinion on the root.
Thus the hard coded relatlive path is a problem.
It is possible to obtain an absolute path at compile time but this relies on
a hack.

criterion will store the files in the workspace root
bench will store the files in the crate root.