`State` gains ` to_static_str()` method (#224)
* feat(common) `State` requires `to_static_str`

This introduces the method `to_static_str` on `State`.

This addresses the need for fast, non-allocating strings from a `State`,
without the risk of method implementation and enum getting out of sync.

Some digging into the efficiency of match statements on enums reveals that
rustc is smart and (under these circumstances) does not search through all
variants.

"Matches on enums with undeclared discriminants (such that the compiler can
choose sequential values) are optimized down to a jump table ..."

[match statement efficiency](https://users.rust-lang.org/t/match-statement-efficiency/4488/2)

* fix(event_body): One mention of 'properties' in clone docs slipped throug

* fix(clippy): Clippy: `#[must_use]` for `to_static_str`

Signed-off-by: Luuk van der Duim <luukvanderduim@gmail.com>

---------

Signed-off-by: Luuk van der Duim <luukvanderduim@gmail.com>