feat(common): zero-copy AT-SPI2 event bodies during user facing event conversion (#230)
To obtain user facing event bodies we went from `serialized body` -> `EventBodyOwned` -> `user facing event`.
Observe that the intermediate `EventBodyOwned` is only short lived and it's scope is limited to the conversion. 

If there ever was a perfect opportunity to avoid instantiating a type, this is it.  
This PR eliminates the instantiation of an owned `EventBodyOwned` in favor of `EventBody<'_>` which can be either borrowed or owned. Which zero-copy de-serializes into its borrowed variant.

We also seized another opportunities to make the path from `Message` -> `Event` more efficient as we realized that the message header was read multiple times in that path. Instead of instantiating it, we pass it along from the first time it is instantiated.

1000 `Message` into user facing events:

|:---------------------:|:-----------------------:|
| main (before) | zero-copy branch |
| 498.24 µs |  248.31 µs |

(Core Ultra 155H)

This PR includes cleanups and refactoring all over the board.