fix (proxies): make `ProxyExt` convert into valid proxies #237
See #237 for description of the problem.

We can not rely on `ThisInterfaceProxy::from` to do a meaningful conversion.
The builder is `async`, so the current solution also became `async` even if
the builder without caching is actually sync - makeing the async wrapping just a noop wrapping.

The solution in this fix no longer stores proxies for later use.

`Option::get_or_insert` required an `FnOnce` closure which, with
our MSRV, cannot be `async`.