fix (proxies): make `ProxyExt` convert into valid proxies #237 (#239)
Fixes: #237

We cannot rely on `ThisInterfaceProxy::from` to do a conversion.

The builder is `async`, so the current solution also became `async` even if
the builder without caching is actually sync - making the `async` wrapping more or less a no-op .

The solution in this fix no longer stores proxies for later use.

`Option::get_or_insert` requires an `FnOnce` closure which, with
our MSRV, cannot be `async`.