refactor(p2p): Make `Peers` it's own type
Introduce `Peers`
It handels mutations on itself, single point of concern.
By keeping the implementation hidden we can later on change the implementation if we see fit.
By keeping mutations to the data itself, this greatly reduces the size of the listener task,
whis makes Clippy and your author happy.