refactor(p2p): rework error handling with tracing
- (!!) Adds `default_path` to `ApplicationProxy` as it is needed.
I did not feel like adding either a const or a magic string slice.
But arguably this needs to be in it's own PR.

- Move taking ownership to where its concern lies, with the methods of `Peers`,
not in the updater task.

- DRY error handling in tracing context in updater task

- Start `Peers updater task` using `zbus::Connection`'s `Executor` to abstracting over
the async runtimes like zbus does.

Note:

Every connection has an `Executor` instance on which tasks may be launched.
If the `Connection::Builder` is left untouched (ie. default) and the application on tokio, (zbus feature "tokio" set) the Executor instance is empty and all zbus' tasks are run on the tokio runtime.
In case of any other executor (smol, glommio, ringbahn etc), by default each Connection will spin up a thread with an `async_executor::Executor`. Generally application will have a single connection, but using `p2p` you will have a connection with each application that supports it, and because we currently build default p2p connections, on anything but tokio will get you an extra thread with an async_executor for each connection!
So picking `smol` won't necessarily make your application small in the context of p2p. Atleast not with how I've integrated p2p currently. Good to know.

 Conflicts:
	atspi/Cargo.toml