doc(connection): fixes and visibility
The `Peers` type is private, an implementation detail to the public
`P2P` trait on `AccessibilityConnection`.

The method pretended to be more public than it's type, which it can't.
Since `fidn_peer` on `P2P` also has an example, and to avoid confusion,
the example on the method is now removed.

Various little fixes elsewhere.