build: depend on true minimal versions
```shell
cargo +nightly generate-lockfile -Z direct-minimal-versions
```

Generates a lockfile for reproducable builds.
Also it tries to unify the dependency graph by minimal version first.

example:
Consider we depend on bar and foo and bar itself also requires foo.
Our toml says we require foo version "1" (^1.0.0) but bar is more
restrictive and requires foo version "1.0.10" then the unified
minimal dependency is "1.0.10". In a sense we overpromised.

This PR corrects for overpromising and adds a unified Cargo.lock.
As the cargo team advised in august 2023.