fix(conn/p2p): Gracefully handle invalid p2p address (#287)
* fix(conn/p2p): Gracefully handle invalid p2p address

Accerciser returns an empty string when `GetApplicationBusAddress`
is called on it.

The fix makes the problem known in debug and when tracing is set,
but stays silent otherwise as it isn't a big deal.

Also:

- fixes a new clippy lint where we must explicitly `.clone()`
a string instead of calling `to_string()` because it hides the
clone.

- fixes a handful of missing `# Example` tags / headers and
other minor brush-ups.