refactor(common): `ObjectRef::Owned` name and path types
The way to achieve a `const` constructor  `ObjectRefOwned` was to
swap out the inner types of `ObjectRef::Owned` from
`UniqueNameOnwed` and `ObjectPathOwned` wrappers to `UniqueName<'static>`
and `ObjectPath<'static>` types which themselves do have a `const`
constructor which their Owned wrapper lacks.

A bunch of methods had to be adjusted accordingly, but it worked out nicely.

Fixes:
- rustdocs on `P2P::object_as_accessible`