fix: Deserializing a null-reference
We receive a binary block of data with an "(so)" signature.

The protocol describes that the combination of an empty string with
a null-path should be considered a null reference.
The protocol does acknowledge that other practices are in use.

Chromium, for instance, uses a name and the null-path.

This commit makes every (so) deserialize to a null-reference if it has a
null path.

Also:
- Adds assertion to ensure no empty names are found in the "happy path".
- Adds null-reference tests.
- Adds `AtspiError::NullRef(&'static str)` to return an error to easily
communicate where the error occured.
- Add warnings on `P2P::object_as_accessible`, `ObjectRefExt::as_accessible / into_accessible`
- Filter or test on null-references in our examples.