`ObjectRef` refactor to address #271, supersedes #272 (#294)
fix: Deserializing a null-reference

We receive a binary block of data with an "(so)" signature.

The AT-SPI protocol describes that the combination of an empty string with
a null-path should be considered a null reference.
The protocol does acknowledge that other practices are in use.

Chromium, for instance, uses a name and the null-path.

This commit makes every (so) deserialize to a null-reference if it has a
null path.