refactor: Add `NonNullObjectRef` as part of `ObjectRef`
We'd like to encode in our APIs that some functions should not be
called with a null-reference. This required us to rethink how we
organize our references.

We landed on:

```rust
enum ObjectRef<'o> {
    Null,
    NonNull(NonNullObjectRef<'o>),
}

enum NonNullObjectRef<'o> {
    Borrowed { name: UniqueName<'o>, path: ObjectPath<'o> },
    Owned { name: OwnedUniqueName, path: OwnedObjectPath }
}

struct ObjectRefOwned(ObjectRef<'static>)
```

Which offers us an `ObjectRef` and `ObjectRefOwned` that can be
deserialized from any reference, including null-referenes. Good for
receiving references.

`NonNullObjectRef` is particularly well suited where we want to
accept object references.