CI: Add git hooks (#303)
This commit adds local git hooks based on those distributed with [zbus](https://github.com/dbus2/zbus).
This adds a "cargo machete" pass to check for unused dependencies to the pre-commit script.

Considering `.git/hooks` is not part of the distribution, contributors are advised to configure git to enable these hooks for this repository:

```sh
git config core.hooksPath .githooks
```

Thank you @C-Loftus for bringing this up and @jonathanhood for the tidy helper script.

Addresses: #298