fix(proxy): Add attributes to fix member re-creation
Zbus' `#[proxy]` macro takes a method or property identifier and tries to recreate the member name from it.
This is certain to fail if the original had two (or more) consecutive capital letters in the member name.

```sh
rg 'property name="[A-Za-z]+"' atspi-common/xml/ -o | rg '[A-Z]{2,}'
```
The problem is prevented in a later version of `zbus_xmlgen`. 
This allowed us to check whether our regex search was exhaustive.

All problematic items now gained an appropriate attribute.
 eg. `#[zbus(name = "NActions"]`

---------

Co-authored-by: Luuk van der Duim <luukvanderduim@gmail.com>